Updated function of links decoration
